chore: merge bagel.py and bagel_umm.py
